+++ application/migrations/004_add_date_format_setting.php
@@ -19,10 +19,13 @@ class Migration_Add_date_format_setting extends EA_Migration {
      */
     public function up()
     {
-        $this->db->insert('settings', [
-            'name' => 'date_format',
-            'value' => 'DMY'
-        ]);
+        if ( ! $this->db->get_where('settings', ['name' => 'date_format'])->num_rows())
+        {
+            $this->db->insert('settings', [
+                'name' => 'date_format',
+                'value' => 'DMY'
+            ]);
+        }
     }
 
     /**
@@ -32,6 +35,9 @@ class Migration_Add_date_format_setting extends EA_Migration {
      */
     public function down()
     {
-        $this->db->delete('settings', ['name' => 'date_format']);
+        if ($this->db->get_where('settings', ['name' => 'date_format'])->num_rows())
+        {
+            $this->db->delete('settings', ['name' => 'date_format']);
+        }
     }
 }
